src: Improve genl family vanished watch handler.
Support the case where the "mptcp" generic netlink family dynamically
vanishes and reappears by moving multicast handler deregistration to
the family_vanished() watch handler.  Such a scenario can happen when
the mptcp_netlink kernel module in the multipath-tcp.org Linux kernel
is compiled as dynamically loadable module instead of one that is
built in to the kernel.