tests: Correct test-start-stop behavior.
The plugin directory used for testing mptcpd start/stop was incorrect,
and triggered an assertion because the wrong plugin was loaded.
Despite the assertion, the `test-start-stop' script did not fail, and
exited successfully instead.

Use the no-op plugin directory for this test since we're not testing
plugin framework internals.  This addresses the assertion.

Make the `test-start-stop' script more robust by leveraging the GNU
`timeout' command to propagate `mptcpd' failures, including
assertions, as well as adding an additional -KILL signal failsafe if
the first -TERM signal didn't trigger `mptcpd' shutdown.