Migrate to multipath-tcp.org generic netlink API. (#15)
* Migrate to multipath-tcp.org generic netlink API.

Replace the deprecated MPTCP generic netlink interface implementation
in mptcpd with the one available in the multipath-tcp.org kernel.
Closes #12.

* src: Improve genl family vanished watch handler.

Support the case where the "mptcp" generic netlink family dynamically
vanishes and reappears by moving multicast handler deregistration to
the family_vanished() watch handler.  Such a scenario can happen when
the mptcp_netlink kernel module in the multipath-tcp.org Linux kernel
is compiled as dynamically loadable module instead of one that is
built in to the kernel.

* tests: Correct test-start-stop behavior.

The plugin directory used for testing mptcpd start/stop was incorrect,
and triggered an assertion because the wrong plugin was loaded.
Despite the assertion, the `test-start-stop' script did not fail, and
exited successfully instead.

Use the no-op plugin directory for this test since we're not testing
plugin framework internals.  This addresses the assertion.

Make the `test-start-stop' script more robust by leveraging the GNU
`timeout' command to propagate `mptcpd' failures, including
assertions, as well as adding an additional -KILL signal failsafe if
the first -TERM signal didn't trigger `mptcpd' shutdown.

* tests: Exercise expected plugin dispatch failure.

Attempt to call a plugin operation with a bad token to verify that no
operation will be dispatched to loaded plugins.