src: Correct token type inconsistency.
Retrieval of the MPTCP connection token associated with a subflow
unnecessarily added another level of indirection with a parameter type
of mptcp_token_t** instead of mptcpd_token_t*.  That caused a
compile-time error in gcc 9.1 and clang 8 builds due to a format
specifier mismatch in a debug log message.  Drop the extra level of
indirection.  Fixes #26.

Reported-By: Mat Martineau <mathew.j.martineau@linux.intel.com>