Revert "src: Correct token type inconsistency."
The original type mptcpd_token_t const** was correct since
l_genl_attr_next() passes a pointer to the attribute back to the
caller, not a copy of the value. The token value was not propagated
back to the functions that called handle_subflow() with the change
being reverted in place.

This reverts commit 3f9b83d0e65cffbc6127455d8122c4983adc18a4.