lib: Make types on both sides of assignment match.
In the mptcpd network monitor, a 32 bit IP address obtained through
the rtnetlink API is of type uint32_t, and is assigned to a temporary
variable of type in_addr_t to better reflect that it is an internet
address.  Both types have the same underlying type, so prioritize
readability and consistency by having the types on either side of the
assignment match.  Go with uint32_t.