Transition to new ELL genl API. (#34)
* src: Transition to new ELL genl API.

The ELL generic netlink API was changed in a non-backward compatible
way in the upcoming 0.21 release.  Update mptcpd to use the new ELL
generic netlink API.

The update slightly modifies behavior so that a l_genl_family object
used for interactions with the kernel cannot be allocated until the
generic netlink family, "mptcp" in this case, appears.  This isn't a
major change, but it will affect some of the mptcpd tests.

* lib: Add a new mptcp_pm_is_ready() function.

Expose a new mptcpd_pm_is_ready() function in the mptcpd library API
that allows a user to know if the path manager is ready for use,
i.e. for interaction with the kernel's generic netlink based path
manager.

* tests: Run PM tests when "mptcp" family appears.

The mptcpd path manager tests are limited without the existence of the
"mptcp" generic netlink family in the kernel.  Only run the tests if
the "mptcp" family is available, and bound the time to wait for it to
appear.  Force the path manager tests to fail if the "mptcp" family
doesn't appear within the allotted time.