src: Fix kernel path manager detection.
The path manager name obtained from
/proc/sys/net/mptcp/mptcp_path_manager was obtained using fgets(),
resulting in the string containing a new line character '\n', which
caused the string comparison to "netlink" to fail.  Use fscanf() to
avoid the new line character, instead.  This corrects a problem where
the user was erroneously warned about the "netlink" path manager not
being selected despite the fact it actually was.