src: Avoid potential family watch/request race.
Add the genl family watch before explicitly requesting the family to
avoid a potential race condition where the family appears after the
request but before the watch.

Do nothing in the family watch/request callback if the necessary
"mptcp" family registration was completed as a result of a previous
call.

Reported-By: Mat Martineau <mathew.j.martineau@linux.intel.com>