lib: Verify plugin op is non-NULL before calling.
While not initially intended, the mptcpd plugin framework allows an
instance of mptcpd_plugin_ops with NULL callback fields to be
registered.  However, the plugin framework did not check if the
callback was NULL prior to calling.  Add the appropriate NULL check
prior to invoking the plugin callback to prevent a seg fault from
occurring.  Fixes #65.