lib: Verify plugin op is non-NULL before calling. (#66)
* tests: Add NULL plugin ops test.

The mptcpd plugin framework currently allows an instance of
mptcpd_plugin_ops with NULL callback fields to be registered.  Add a
test that verifies such a case is handled gracefully.

* lib: Verify plugin op is non-NULL before calling.

While not initially intended, the mptcpd plugin framework allows an
instance of mptcpd_plugin_ops with NULL callback fields to be
registered.  However, the plugin framework did not check if the
callback was NULL prior to calling.  Add the appropriate NULL check
prior to invoking the plugin callback to prevent a seg fault from
occurring.  This also allows the user to avoid having to implement
stubs for unused plugin operations. Fixes #65.