Use kernel <linux/mptcp.h> if available.
Detect and use the kernel <linux/mptcp.h> header if available.  Prefer
it over the local mptcpd copy of the header to help ensure mptcpd
always uses the latest version of the header.  The server-oriented
MPTCP path management generic netlink API is also detected In
preparation for mptcpd's support of that API.