Propagate PM command error to callbacks.
Rather than silently fail and not call the user supplied callback for
the get_addr, dump_addrs, and get_limits path manager commands when a
generic netlink occurs, pass NULL/zero result arguments to the user
callback to at least make the user aware of a problem.  This also
allows the test-commands unit test to fail to reflect the generic
netlink problem.