Fix build regression against ELL 0.31. (#133)
* Fix build regression against ELL 0.33.

The mptcpd ID manager uses the ELL l_hashmap_replace() function, but
that function was introduced in ELL 0.33.  Implement the
l_hashmap_replace() functionality in mptcpd through calls to
l_hashmap_remove() and l_hashmap_insert() when l_hashmap_replace() is
not available.

* Fix build regression against ELL 0.31.

Check if the l_genl_msg_get_extended_error() function is available.
It was introduced in ELL 0.31.  Set the extended error message to NULL
if the function is not available.