addr_adv: keep limit consistent with current addess set
To really use the configured endpoints, the upstream kernel
need suitable limits - e.g. max_subflows must > 0 to create
at least an additional subflows.

Keep track of the number of discovered address, and set the
limits accordingly. Additionally, allows accepting address
notification if we are supposely on the client side.

This change is a no-op when running on top of mptcp.org kernel.