fix failing assertion in test-command
If the current user is not root, the netlink set operation
will fail on upstream kernel, thus the get_limits() command will
read the default value instead of the ones provided by set_limits().

Address the issues explicitly checking the current users and
matching for default if not root.

Additionally update the test-suite, so the no failures are now
expected