Improve addr_adv path manager (#151)
* configuration: add support for addr-flags param

Will be used by the upstream pm netlink in the next
patch. Parse a comma separated token list into the corresponding
address flags and vice versa

* addr_adv: use the config-specified address flags, if any

Additionally we need to expose the configuration struct in
the private header to allow the plugins to access it

* addr_adv: keep limit consistent with current address set

To really use the configured endpoints, the upstream kernel
needs suitable limits - e.g. max_subflows must be > 0 to
create at least additional subflows.

Keep track of the number of discovered address, and set the
limits accordingly. Additionally, allows accepting address
notification if we are supposely on the client side.

This change is a no-op when running on top of mptcp.org kernel.

* configuration: add support for (address) notification flags param

Such parameter will control which addresses will be notified to
the plugins. The available flags are:

existing
   Notify even the addresses existing at startup time

skip_link_local
   Ignore (do not notify) [ipv6] link local addresses

skip_loopback
   Ignore (do not notify) host (loopback) addresses

Will be used by the next patch

* Implement support for notification flags

Notify existing addresses and/or link local one according to
the configured notify flags

* fix test-network-monitor crash

The mentioned test uses a dummy user_data argument, while
the network monitor expect a valid mptcp_pm struct - used
to fetch the relevant config.

Solve the issue removing the latter assumption, and providing
the relevant configuration info at nm creation time as
mptcpd_nm_create() argument

* fix failing assertion in test-command

If the current user is not root, the netlink set operation
will fail on upstream kernel, thus the get_limits() command will
read the default value instead of the ones provided by set_limits().

Address the issues explicitly checking the current users and
matching for default if not root.

Additionally update the test-suite, so the no failures are now
expected

* man: Update mptcpd(8) man page.

Add descriptions of the newly added --addr-flags and --notify-flags
command line options.  Add the ip-mptcp(8) to the "SEE ALSO" section.

* etc: Update system mptcpd configuration file.

Add descriptions of the new "addr-flags" and "notify-flags" mptcpd
configuration file settings.

* Document new notification flags.

Add documentation for the new notification flags to the mptcpd(8) man
page and the mptcpd.conf system configuration file based on Paolo's
earlier commit summary.

Co-authored-by: Paolo Abeni <pabeni@redhat.com>
Co-authored-by: Ossama Othman <ossama.othman@intel.com>