include: Split in-kernel PM to new namespace.
Split path management commands specific to the in-kernel path manager
found in the upstream Linux kernel into a separate "mptcpd_kpm"
code namespace to differentiate ADD_ADDR and REMOVE_ADDR related
operations between the in-kernel and userspace cases.  The generic
netlink commands for the two are different and have different use
cases.