sspi: Update mptcpd_pm_add_addr() call.
The "flags" and "index" parameters for the mptcpd_pm_add_addr()
function were removed in the in-kernel and userspace PM API split.
Drop these arguments from the mptcpd_pm_add_addr() call in the sspi
plugin since they are only needed for the in-kernel (server-oriented)
path management case.  The sspi plugin is meant for use with the
userspace (client-oriented) path management API currently only found
in the multipath-tcp.org kernel.