Split in-kernel PM add_addr/remove_addr commands. (#153)
* include: Split in-kernel PM to new namespace.

Split path management commands specific to the in-kernel path manager
found in the upstream Linux kernel into a separate "mptcpd_kpm"
code namespace to differentiate ADD_ADDR and REMOVE_ADDR related
operations between the in-kernel and userspace cases.  The generic
netlink commands for the two are different and have different use
cases.

* Support split PM APIs in command implementations.

* Use new split command API in test-commands unit test.

* addr_adv: Use new mptcpd_kpm_*() commands.

The "addr_adv" is meant for use with the in-kernel path manager.  Use
the newly "mptcpd_kpm_*()" family of PM command functions
accordingly.

* sspi: Update mptcpd_pm_add_addr() call.

The "flags" and "index" parameters for the mptcpd_pm_add_addr()
function were removed in the in-kernel and userspace PM API split.
Drop these arguments from the mptcpd_pm_add_addr() call in the sspi
plugin since they are only needed for the in-kernel (server-oriented)
path management case.  The sspi plugin is meant for use with the
userspace (client-oriented) path management API currently only found
in the multipath-tcp.org kernel.

* include: Remove obviated function pointers.

Remove callback function pointers from the mptcpd_pm_cmd_ops structure
that were left over from the earlier move of the function pointers
specific to the in-kernel path manager to the new mptcpd_kpm_cmd_ops
structure.