m4: Package latest AX_CODE_COVERAGE macro source.
The version of the AX_CODE_COVERAGE macro shipped with the version of
Ubuntu found in the GitHub Actions environment has a make variable
assignment problem that introduces a recursion related problem when
running "./configure --enable-code-coverage; make distcheck".  Package
the latest version of that macro with mptcpd to address that problem.