tests: Handle case MPTCP IDs exist in the kernel.
The current test-commands unit test assumes that it can use hard-coded
MPTCP address ID test values, but that could fail if the same address
ID is currently in use by the in-kernel path manager.  Leverage the
mptcpd global MPTCP address ID manager to generate an ID suitable for
the test since it attempts to stay in sync with the kernel.