tests: Setup test address before command test run.
The mptcpd_kpm_add_addr() and mptcpd_kpm_get_addr() command tests will
fail if the test IP address is not bound to a network interface.  Set
up the test IP address on the loopback interface prior to the test
run, and tear it down afterward.  This requires that the test be run
with CAP_NET_ADMIN capability, such as running it through the root
user.

The previous behavior of ignoring and logging failures due to lack of
the CAP_NET_ADMIN capability is still preserved.