tests: Fix test-commands test case failures. (#172)
With these changes in place the full set of test cases in the
test-commands unit test now pass when run as root or with the
CAP_NET_ADMIN capability. Test cases that require CAP_NET_ADMIN
but run as an unprivileged user still fail as expected. Failures
in that cases are ignored.

The test suite was run with the following set of commands:

    Unprivileged user
        make check
    root user
        sudo make check
    Unprivileged user with CAP_NET_ADMIN capability
      sudo capsh \
        --caps="cap_net_admin+eip cap_setpcap,cap_setuid,cap_setgid+ep" \
        --keep=1 --user=$USER --addamb=cap_net_admin -- -c "make check"

There is one caveat to keep in mind when running make check or
just the test-commands unit test after a privileged test run: the
get_limits test case assumes that the subflows and
add_addr_accepted MPTCP limits are always zero when not run as
root, which is not true after the privileged test run, resulting
in failure. MPTCP limits should be reset to zero between
privileged and unprivileged test runs, e.g.:

$ sudo make check
$ sudo ip mptcp limits set subflows 0 add_addr_accepted 0
$ make check

Relevant commits:
* tests: Handle case MPTCP IDs exist in the kernel.

The current test-commands unit test assumes that it can use hard-coded
MPTCP address ID test values, but that could fail if the same address
ID is currently in use by the in-kernel path manager.  Leverage the
mptcpd global MPTCP address ID manager to generate an ID suitable for
the test since it attempts to stay in sync with the kernel.

* tests: Leverage mptcpd pm_ready callback.

Instead of relying on the appearance of the MPTCP generic netlink path
manager family, use the mptcpd path manager "pm_ready" callback
feature to determine when to run the test cases in the test-commands
unit test.  This ensures that mptcpd is fully initialized prior to
running the test cases.

* tests: Remove unnecessary mptcpd_pm_ready() call.

The mptcpd_pm_ready() call in the test-commands unit test is no longer
necessary since the tests now are only run when the mptcpd path
manager is guaranteed to be fully initialized and ready.

* tests: Call flush_addrs last.

Call flush_addrs after all kernel path management command test cases
are run to make sure resources, such as MPTCP address/IP mappings,
used between between test cases are valid.  Calling flush_addrs wipes
out all such mappings.

* tests: Setup test address before command test run.

The mptcpd_kpm_add_addr() and mptcpd_kpm_get_addr() command tests will
fail if the test IP address is not bound to a network interface.  Set
up the test IP address on the loopback interface prior to the test
run, and tear it down afterward.  This requires that the test be run
with CAP_NET_ADMIN capability, such as running it through the root
user.

The previous behavior of ignoring and logging failures due to lack of
the CAP_NET_ADMIN capability is still preserved.

Co-authored-by: Mat Martineau <mathew.j.martineau@linux.intel.com>