configure: Avoid undefining _FORTIFY_SOURCE.
The configure script ended up causing _FORTIFY_SOURCE to be undefined
if the user explicitly adds _FORTIFY_SOURCE=2 to the CPPFLAGS.  For
example, running the configure script like so:

    CPPFLAGS=-D_FORTIFY_SOURCE=2 ./configure

ultimately caused CPPFLAGS to contain "-D_FORTIFY_SOURCE=2
-U_FORTIFY_SOURCE" instead of "-D_FORTIFY_SOURCE=2
-U_FORTIFY_SOURCE -D_FORTIFY_SOURCE=2" due to the AX_APPEND_FLAG()
autoconf archive macro behavior of not appending duplicate flags.

Only append -D_FORTIFY_SOURCE=2 to CPPFLAGS if the user hasn't defined
_FORTIFY_SOURCE rather than attempt to override the user provided
value.  This is more in line with Autoconf conventions as well.