configure: Automatically choose default plugin. (#205)
Choose the default mptcpd plugin based on the kernel mptcpd is
configured to support.  Specifically, the addr_adv plugin will
continue to be the default for the upstream kernel, and the sspi
plugin will be the default for the multipath-tcp.org kernel.  The
MPTCP path management generic netlink API used by addr_adv is not
supported by the multipath-tcp.org kernel, meaning the addr_adv plugin
is not a suitable default for it.