Skip some tests on non-MPTCP kernels. (#215)
* tests: Add MPTCP kernel check C function.

* tests: Skip some tests on non-MPTCP kernels.

The mptcpd unit tests test-commands, test-path-manager and
test-mptcpwrap require a MPTCP capable kernel to run.  Make those
tests return with an exit status (77) that causes the Automake test
driver script, i.e. `test-driver' to treat the tests as skipped.  This
allows the test suite to succeed when MPTCP is not available in the
kernel.

The test summary from a `make check' run would look similar to the
following when run on a non-MPTCP capable kernel:

 ...
 PASS: test-network-monitor
 SKIP: test-path-manager
 SKIP: test-commands
 PASS: test-configuration
 ...
 SKIP: test-mptcpwrap
 =========================================================================
 Testsuite summary for mptcpd 0.9
 =========================================================================
 # TOTAL: 17
 # PASS:  14
 # SKIP:  3
 # XFAIL: 0
 # FAIL:  0
 # XPASS: 0
 # ERROR: 0
 =========================================================================

Fixes #204.