Fix compatibility with ELL 0.45 and later (#233)
ELL 0.45 and later have header files that use a trailing semicolon with
the DEFINE_CLEANUP_FUNC() macro, which is not compatible with the
"-pedantic" flag in gcc. This causes the mptcpd code to issue the
following error:

"ISO C does not allow extra ‘;’ outside of a function [-Werror=pedantic]"

Wrap the ELL includes in some pragmas to avoid this error.