configuration: Fix use-after-free in merged configuration for load-plugins (#245)
The section of merge_config() that copies entries from one
plugins_to_load queue to another was copying string pointers from one to
the other. When the source queue was freed, the string pointers in the
destination queue became invalid.

This manifested as garbage strings for the plugins_to_load value if the
load-plugins line was used in mptcpd.conf

To fix, populate the destination queue entries with l_strdup()'d
strings.