hash_sockaddr: Include IP port as part of the key.
Hash the IP port as part of the key to ensure the resulting key hash
is different between sockaddrs with the same IP address but different
ports.

This is necessary because the mptcpd listener manager would otherwise
only create listeners for a given IP address once even if the port
differs in subsequent mptcpd_lm_listen() calls.