listener_manager: Track ephemeral ports.
Track the kernel assigned ephemeral port if the user provided a
sockaddr with a zero port to mptcpd_lm_listen(), and return that port
to the user.  The user is expected to pass a sockaddr with the port
returned from mptcpd_lm_listen() in a subsequent call to
mptcpd_lm_close().  Passing a sockaddr with a zero port to
mptcpd_lm_close() will result in failure since the listener manager
now only tracks listeners with non-zero addresses.