scripts: get debug info for bug reports
This is a simple script that is here to get debug info. It will be
helpful to get more details in case of issues with MPTCP in general,
not specific to mptcpd or mptcpize then.

The script is basic, supporting /bin/sh with a minimum of dependences.

It should be used with '--collect' or '--reproduce', as shown by the
help menu.

Please note that a lot of efforts have been done recently to improve the
documentation on mptcp.dev website, on the kernel side, and on IPRoute
side. Thanks to that, this script can be small, focussing on this small
set of commands for the moment. It can always be improved later on.

The idea is to place it here in mptcpd, to have it somewhere, and
ideally already installed on some machines, in /usr/libexec dir, ready
to be used without having to install it.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>