git: ignore hidden files / directories
Similar to what is done in the Linux kernel: very likely, hidden files
and directories don't need to be tracked in git. There are exceptions
of course, like the '.github' directory, but we can always force their
inclusions with 'git add -f'.

This will at least ignore the '.cache' directory created by clangd.
Editors tend to create hidden directories and files too, so that would
cover that too.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>