gh: daily validation of the ELL compatibility
We recently had some issues because the ELL API has been broken without
notice.

Simply adding a cron to run the tests once a day during the week should
tell us when an incompatible change has been introduced on ELL side, so
we can get a bit of time to prepare a fix and a new release.

Link: #302
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>