tests: commands: only use one subtest (#322)
ELL 0.72 recently changed the way the subtests are executed: each one is
now executed in a dedicated process, forked from the same parent [1].

This is causing issues for the 'commands' test, because some shared info
are shared between subtests. 'info' could be created in a shared memory
with mmap(), but it looks like there are more changes needed in mptcpd
itself to cope with that, probably the memory allocated for the pm
structure and more.

Then, for the moment, it looks better and easier to simplify things, an
only use one subtest.

Link: https://git.kernel.org/pub/scm/libs/ell/ell.git/commit/?id=e0628c4 [1]

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>