interface: add 'deny_join_id0' parameter
This flag being added on the kernel side [1]. It is set when the other
peer requested not to create new subflows to the initial IP address and
port.

The 'new_connection' and 'connection_established' interfaces are then
modified: a new deny_join_id0 parameter is added. LIB_CURRENT, the sspi
plugin, and different tests have been modified accordingly.

Link: https://lore.kernel.org/20250909-mptcp-pm-user-c-flag-v2-0-a6f9542481c5@kernel.org
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>