interface: add `deny_join_id0` parameter (#323)
* interface: add 'deny_join_id0' parameter

This flag being added on the kernel side [1]. It is set when the other
peer requested not to create new subflows to the initial IP address and
port.

The 'new_connection' and 'connection_established' interfaces are then
modified: a new deny_join_id0 parameter is added. LIB_CURRENT, the sspi
plugin, and different tests have been modified accordingly.

Link: https://lore.kernel.org/20250909-mptcp-pm-user-c-flag-v2-0-a6f9542481c5@kernel.org
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>

* events: support server-side flag

This flag is being added in the kernel [1].

It is going to replace the server-side attribute. Check both to be
future proof.

Link: https://lore.kernel.org/20250909-mptcp-pm-user-server-side-flag-v1-0-cb8e2b8d1c0c@kernel.org [1]
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>

---------

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>