github: ell: clone from Google servers and retry
Recently, we had two jobs in a row that failed to clone ELL repo.
Probably because of the restrictions in place at git.kernel.org, and the
fact two clone are being done in parallel, likely from the same IP
range.

Switching to googlesource.com, which should be more tolerant, and retry
max 30 times with a 30 seconds sleep in between.

If it is really not possible to clone the repo, the 'ell' dir will not
be created.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>