plugin: expose subflow error to plugins
Add error parameter to subflow_closed callback to allow plugins
to identify the reason for subflow closure, such as network errors.

The error value corresponds to the MPTCP_ATTR_ERROR attribute from
the kernel, which equals sk_err from struct sock. If the error
attribute is not present in the netlink message, it defaults to 0.

Fixes #330